Language.LSP.MetaModel.Types

Description

This defines the types of the LSP "metamodel", which is a machine-readable format specifying the types used in the LSP protocol.

The type system is quite typescript-y, which isn't surprising given that the whole protocol is very typescript-y.

A typescript version of the metamodel types can be found here, which is useful for constructing this Haskell model of them: https:/github.commicrosoftvscode-languageserver-nodeblobmaintoolssrcmetaModel.ts

Documentation

data MessageDirection

What direction is this message sent in: server to client, client to server, or both?

Constructors

ServerToClient

ClientToServer

Both

data BaseTypeName

The "base types" in the metamodel.

Constructors

URI

DocumentUri

Integer

UInteger

Decimal

RegExp

String

Boolean

Null

data Property

A property of a structure.

Constructors

Property

Fields

name :: Text

type_ :: Type

optional :: Maybe Bool

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data StructureLiteral

An anonymous structure type.

Constructors

StructureLiteral

Fields

properties :: [Property]

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data Type

The various kinds of type in the metamodel.

Constructors

BaseType

Fields

btName :: BaseTypeName

ReferenceType

Fields

rtName :: Text

ArrayType

Fields

atElement :: Type

MapType

Fields

mKey :: Type

mValue :: Type

AndType

Fields

aItems :: NonEmpty Type

OrType

Fields

oItems :: NonEmpty Type

TupleType

Fields

tItems :: [Type]

StructureLiteralType

Fields

stlValue :: StructureLiteral

StringLiteralType

Fields

slValue :: Text

IntegerLiteralType

Fields

ilValue :: Integer

BooleanLiteralType

Fields

blValue :: Bool

data Request

A request message.

Constructors

Request

Fields

method :: Text

params :: Maybe Type

result :: Type

partialResult :: Maybe Type

errorData :: Maybe Type

registrationOptions :: Maybe Type

messageDirection :: MessageDirection

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data Notification

A notification message.

Constructors

Notification

Fields

method :: Text

params :: Maybe Type

registrationOptions :: Maybe Type

messageDirection :: MessageDirection

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data Structure

A structure type.

Constructors

Structure

Fields

name :: Text

extends :: Maybe [Type]

mixins :: Maybe [Type]

properties :: [Property]

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data TypeAlias

A type alias.

Constructors

TypeAlias

Fields

name :: Text

type_ :: Type

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data TextOrInteger

This is just 'string | int' on the typescript side, but it's convenient to have a proper type here.

Constructors

T Text

I Integer

data EnumerationEntry

An entry in an enumeration.

Constructors

EnumerationEntry

Fields

name :: Text

value :: TextOrInteger

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data Enumeration

An enumeration type.

Constructors

Enumeration

Fields

name :: Text

type_ :: Type

values :: [EnumerationEntry]

supportsCustomValues :: Maybe Bool

documentation :: Maybe Text

since :: Maybe Text

proposed :: Maybe Bool

deprecated :: Maybe Text

data MetaData

Metadata about the metamodel iteslf.

Constructors

MetaData

Fields

version :: Text

data MetaModel

The entire metamodel.

Constructors

MetaModel

Fields

metaData :: MetaData

requests :: [Request]

notifications :: [Notification]

structures :: [Structure]

enumerations :: [Enumeration]

typeAliases :: [TypeAlias]

loadMetaModelFromFile :: FilePath -> Q Exp