Language.LSP.Protocol.Capabilities

Documentation

data LSPVersion

A specific version of the LSP specification.

Constructors

LSPVersion Int Int

Client capabilities

type family ClientCapability (m :: Method f t) where ...

The client capability associated with a given method.

Where several methods are provided together (e.g. the three callHierarchy methods), we associate all of them with the capaiblity, even if there is one "primary" method.

Equations

ClientCapability 'Method_TextDocumentDeclaration = DeclarationClientCapabilities

ClientCapability 'Method_TextDocumentImplementation = ImplementationClientCapabilities

ClientCapability 'Method_TextDocumentTypeDefinition = TypeDefinitionClientCapabilities

ClientCapability 'Method_TextDocumentHover = HoverClientCapabilities

ClientCapability 'Method_TextDocumentSignatureHelp = SignatureHelpClientCapabilities

ClientCapability 'Method_TextDocumentDefinition = DefinitionClientCapabilities

ClientCapability 'Method_TextDocumentReferences = ReferenceClientCapabilities

ClientCapability 'Method_TextDocumentDocumentHighlight = DocumentHighlightClientCapabilities

ClientCapability 'Method_TextDocumentDocumentSymbol = DocumentSymbolClientCapabilities

ClientCapability 'Method_TextDocumentFoldingRange = FoldingRangeClientCapabilities

ClientCapability 'Method_TextDocumentSelectionRange = SelectionRangeClientCapabilities

ClientCapability 'Method_WorkspaceExecuteCommand = ExecuteCommandClientCapabilities

ClientCapability 'Method_TextDocumentLinkedEditingRange = LinkedEditingRangeClientCapabilities

ClientCapability 'Method_TextDocumentMoniker = MonikerClientCapabilities

ClientCapability 'Method_WorkspaceWorkspaceFolders = Bool

ClientCapability 'Method_TextDocumentCompletion = CompletionClientCapabilities

ClientCapability 'Method_CompletionItemResolve = CompletionClientCapabilities

ClientCapability 'Method_TextDocumentCodeAction = CodeActionClientCapabilities

ClientCapability 'Method_CodeActionResolve = CodeActionClientCapabilities

ClientCapability 'Method_TextDocumentCodeLens = CodeLensClientCapabilities

ClientCapability 'Method_CodeLensResolve = CodeLensClientCapabilities

ClientCapability 'Method_WorkspaceCodeLensRefresh = CodeLensWorkspaceClientCapabilities

ClientCapability 'Method_TextDocumentDocumentLink = DocumentLinkClientCapabilities

ClientCapability 'Method_DocumentLinkResolve = DocumentLinkClientCapabilities

ClientCapability 'Method_WorkspaceSymbol = WorkspaceSymbolClientCapabilities

ClientCapability 'Method_WorkspaceSymbolResolve = WorkspaceSymbolClientCapabilities

ClientCapability 'Method_TextDocumentRename = RenameClientCapabilities

ClientCapability 'Method_TextDocumentPrepareRename = RenameClientCapabilities

ClientCapability 'Method_TextDocumentDocumentColor = DocumentColorClientCapabilities

ClientCapability 'Method_TextDocumentColorPresentation = DocumentColorClientCapabilities

ClientCapability 'Method_TextDocumentFormatting = DocumentFormattingClientCapabilities

ClientCapability 'Method_TextDocumentRangeFormatting = DocumentRangeFormattingClientCapabilities

ClientCapability 'Method_TextDocumentOnTypeFormatting = DocumentOnTypeFormattingClientCapabilities

ClientCapability 'Method_TextDocumentPrepareCallHierarchy = CallHierarchyClientCapabilities

ClientCapability 'Method_CallHierarchyIncomingCalls = CallHierarchyClientCapabilities

ClientCapability 'Method_CallHierarchyOutgoingCalls = CallHierarchyClientCapabilities

ClientCapability 'Method_TextDocumentSemanticTokensFull = SemanticTokensClientCapabilities

ClientCapability 'Method_TextDocumentSemanticTokensFullDelta = SemanticTokensClientCapabilities

ClientCapability 'Method_TextDocumentSemanticTokensRange = SemanticTokensClientCapabilities

ClientCapability 'Method_WorkspaceSemanticTokensRefresh = SemanticTokensWorkspaceClientCapabilities

ClientCapability 'Method_TextDocumentPrepareTypeHierarchy = TypeHierarchyClientCapabilities

ClientCapability 'Method_TypeHierarchySupertypes = TypeHierarchyClientCapabilities

ClientCapability 'Method_TypeHierarchySubtypes = TypeHierarchyClientCapabilities

ClientCapability 'Method_TextDocumentInlineValue = InlineValueClientCapabilities

ClientCapability 'Method_WorkspaceInlineValueRefresh = InlineValueWorkspaceClientCapabilities

ClientCapability 'Method_TextDocumentInlayHint = InlayHintClientCapabilities

ClientCapability 'Method_InlayHintResolve = InlayHintClientCapabilities

ClientCapability 'Method_WorkspaceInlayHintRefresh = InlayHintWorkspaceClientCapabilities

ClientCapability 'Method_TextDocumentDiagnostic = DiagnosticClientCapabilities

ClientCapability 'Method_WorkspaceDiagnostic = DiagnosticWorkspaceClientCapabilities

ClientCapability 'Method_WorkspaceDiagnosticRefresh = DiagnosticWorkspaceClientCapabilities

ClientCapability 'Method_WorkspaceWillCreateFiles = FileOperationClientCapabilities

ClientCapability 'Method_WorkspaceWillRenameFiles = FileOperationClientCapabilities

ClientCapability 'Method_WorkspaceWillDeleteFiles = FileOperationClientCapabilities

ClientCapability 'Method_WorkspaceDidCreateFiles = FileOperationClientCapabilities

ClientCapability 'Method_WorkspaceDidRenameFiles = FileOperationClientCapabilities

ClientCapability 'Method_WorkspaceDidDeleteFiles = FileOperationClientCapabilities

ClientCapability 'Method_TextDocumentDidOpen = TextDocumentSyncClientCapabilities

ClientCapability 'Method_TextDocumentDidChange = TextDocumentSyncClientCapabilities

ClientCapability 'Method_TextDocumentDidClose = TextDocumentSyncClientCapabilities

ClientCapability 'Method_TextDocumentDidSave = TextDocumentSyncClientCapabilities

ClientCapability 'Method_TextDocumentWillSave = TextDocumentSyncClientCapabilities

ClientCapability 'Method_TextDocumentWillSaveWaitUntil = TextDocumentSyncClientCapabilities

ClientCapability 'Method_NotebookDocumentDidOpen = NotebookDocumentSyncClientCapabilities

ClientCapability 'Method_NotebookDocumentDidChange = NotebookDocumentSyncClientCapabilities

ClientCapability 'Method_NotebookDocumentDidSave = NotebookDocumentSyncClientCapabilities

ClientCapability 'Method_NotebookDocumentDidClose = NotebookDocumentSyncClientCapabilities

ClientCapability 'Method_WorkspaceDidChangeConfiguration = DidChangeConfigurationClientCapabilities

ClientCapability 'Method_WorkspaceDidChangeWatchedFiles = DidChangeWatchedFilesClientCapabilities

ClientCapability 'Method_WorkspaceDidChangeWorkspaceFolders = Void

ClientCapability 'Method_TextDocumentPublishDiagnostics = PublishDiagnosticsClientCapabilities

ClientCapability 'Method_WorkspaceConfiguration = Bool

ClientCapability 'Method_WorkspaceApplyEdit = Bool

ClientCapability 'Method_WindowWorkDoneProgressCreate = Bool

ClientCapability 'Method_WindowWorkDoneProgressCancel = Bool

ClientCapability 'Method_WindowShowMessage = ShowMessageRequestClientCapabilities

ClientCapability 'Method_WindowShowMessageRequest = ShowMessageRequestClientCapabilities

ClientCapability 'Method_WindowShowDocument = ShowDocumentClientCapabilities

ClientCapability ('Method_Progress :: Method f 'Notification) = Void

ClientCapability 'Method_WindowLogMessage = Void

ClientCapability 'Method_ClientRegisterCapability = Void

ClientCapability 'Method_ClientUnregisterCapability = Void

ClientCapability 'Method_Initialize = Void

ClientCapability 'Method_Initialized = Void

ClientCapability 'Method_Shutdown = Void

ClientCapability 'Method_Exit = Void

ClientCapability 'Method_TelemetryEvent = Void

ClientCapability 'Method_SetTrace = Void

ClientCapability 'Method_LogTrace = Void

ClientCapability ('Method_CancelRequest :: Method f 'Notification) = Void

ClientCapability ('Method_CustomMethod s :: Method f t) = Void

clientCapability :: forall {f :: MessageDirection} {t :: MessageKind} (m :: Method f t). SMethod m -> Lens' ClientCapabilities (Maybe (ClientCapability m))

A lens which focusses on the (possibly absent) client capability associated with a method.

fullLatestClientCaps :: ClientCapabilities

Client capabilities for full support of the current LSP specification.

fullClientCapsForVersion :: LSPVersion -> ClientCapabilities

Client capabilities for full support of the LSP specification up until a version.

fullClientCapsForVersionAndMethod :: forall {f :: MessageDirection} {t :: MessageKind} (m :: Method f t). LSPVersion -> SMethod m -> Maybe (ClientCapability m)

Client capabilities for full support of a specific method in the LSP specification up until a version.

dynamicRegistrationSupported :: forall {f :: MessageDirection} {t :: MessageKind} (m :: Method f t). SMethod m -> ClientCapabilities -> Bool

Whether the client supports dynamic registration for the given method.

Note that here we only consider the "main" method against which you dynamically register, so even though e.g. we associate the client capabilities for code lenses with `codeLens/resolve`, we don't ever say that you can dynamically register `codeLens/resolve`, because you in fact need to register `textDocument/codeLens`.

Server capabilities

type family ServerCapability (m :: Method f t) where ...

The server capability associated with a given method.

Where several methods are provided together (e.g. the three callHierarchy methods), we associate all of them with the capaiblity, even if there is one "primary" method.

For methods which strictly only need a client capability but which are closely related to a server capability (e.g. `codeLens/refresh`), we also associate them with that server capability.

Equations

ServerCapability 'Method_TextDocumentDeclaration = Bool |? (DeclarationOptions |? DeclarationRegistrationOptions)

ServerCapability 'Method_TextDocumentImplementation = Bool |? (ImplementationOptions |? ImplementationRegistrationOptions)

ServerCapability 'Method_TextDocumentTypeDefinition = Bool |? (TypeDefinitionOptions |? TypeDefinitionRegistrationOptions)

ServerCapability 'Method_TextDocumentHover = Bool |? HoverOptions

ServerCapability 'Method_TextDocumentSignatureHelp = SignatureHelpOptions

ServerCapability 'Method_TextDocumentDefinition = Bool |? DefinitionOptions

ServerCapability 'Method_TextDocumentReferences = Bool |? ReferenceOptions

ServerCapability 'Method_TextDocumentDocumentHighlight = Bool |? DocumentHighlightOptions

ServerCapability 'Method_TextDocumentDocumentSymbol = Bool |? DocumentSymbolOptions

ServerCapability 'Method_TextDocumentFoldingRange = Bool |? (FoldingRangeOptions |? FoldingRangeRegistrationOptions)

ServerCapability 'Method_TextDocumentSelectionRange = Bool |? (SelectionRangeOptions |? SelectionRangeRegistrationOptions)

ServerCapability 'Method_WorkspaceExecuteCommand = ExecuteCommandOptions

ServerCapability 'Method_TextDocumentLinkedEditingRange = Bool |? (LinkedEditingRangeOptions |? LinkedEditingRangeRegistrationOptions)

ServerCapability 'Method_TextDocumentMoniker = Bool |? (MonikerOptions |? MonikerRegistrationOptions)

ServerCapability 'Method_WorkspaceWorkspaceFolders = WorkspaceFoldersServerCapabilities

ServerCapability 'Method_TextDocumentCompletion = CompletionOptions

ServerCapability 'Method_CompletionItemResolve = CompletionOptions

ServerCapability 'Method_TextDocumentCodeAction = Bool |? CodeActionOptions

ServerCapability 'Method_CodeActionResolve = Bool |? CodeActionOptions

ServerCapability 'Method_TextDocumentCodeLens = CodeLensOptions

ServerCapability 'Method_CodeLensResolve = CodeLensOptions

ServerCapability 'Method_WorkspaceCodeLensRefresh = CodeLensOptions

ServerCapability 'Method_TextDocumentDocumentLink = DocumentLinkOptions

ServerCapability 'Method_DocumentLinkResolve = DocumentLinkOptions

ServerCapability 'Method_WorkspaceSymbol = Bool |? WorkspaceSymbolOptions

ServerCapability 'Method_WorkspaceSymbolResolve = Bool |? WorkspaceSymbolOptions

ServerCapability 'Method_TextDocumentRename = Bool |? RenameOptions

ServerCapability 'Method_TextDocumentPrepareRename = Bool |? RenameOptions

ServerCapability 'Method_TextDocumentDocumentColor = Bool |? (DocumentColorOptions |? DocumentColorRegistrationOptions)

ServerCapability 'Method_TextDocumentColorPresentation = Bool |? (DocumentColorOptions |? DocumentColorRegistrationOptions)

ServerCapability 'Method_TextDocumentFormatting = Bool |? DocumentFormattingOptions

ServerCapability 'Method_TextDocumentRangeFormatting = Bool |? DocumentRangeFormattingOptions

ServerCapability 'Method_TextDocumentOnTypeFormatting = DocumentOnTypeFormattingOptions

ServerCapability 'Method_TextDocumentPrepareCallHierarchy = Bool |? (CallHierarchyOptions |? CallHierarchyRegistrationOptions)

ServerCapability 'Method_CallHierarchyIncomingCalls = Bool |? (CallHierarchyOptions |? CallHierarchyRegistrationOptions)

ServerCapability 'Method_CallHierarchyOutgoingCalls = Bool |? (CallHierarchyOptions |? CallHierarchyRegistrationOptions)

ServerCapability 'Method_TextDocumentSemanticTokensFull = SemanticTokensOptions |? SemanticTokensRegistrationOptions

ServerCapability 'Method_TextDocumentSemanticTokensFullDelta = SemanticTokensOptions |? SemanticTokensRegistrationOptions

ServerCapability 'Method_TextDocumentSemanticTokensRange = SemanticTokensOptions |? SemanticTokensRegistrationOptions

ServerCapability 'Method_WorkspaceSemanticTokensRefresh = SemanticTokensOptions |? SemanticTokensRegistrationOptions

ServerCapability 'Method_TextDocumentPrepareTypeHierarchy = Bool |? (TypeHierarchyOptions |? TypeHierarchyRegistrationOptions)

ServerCapability 'Method_TypeHierarchySupertypes = Bool |? (TypeHierarchyOptions |? TypeHierarchyRegistrationOptions)

ServerCapability 'Method_TypeHierarchySubtypes = Bool |? (TypeHierarchyOptions |? TypeHierarchyRegistrationOptions)

ServerCapability 'Method_TextDocumentInlineValue = Bool |? (InlineValueOptions |? InlineValueRegistrationOptions)

ServerCapability 'Method_WorkspaceInlineValueRefresh = Bool |? (InlineValueOptions |? InlineValueRegistrationOptions)

ServerCapability 'Method_TextDocumentInlayHint = Bool |? (InlayHintOptions |? InlayHintRegistrationOptions)

ServerCapability 'Method_InlayHintResolve = Bool |? (InlayHintOptions |? InlayHintRegistrationOptions)

ServerCapability 'Method_WorkspaceInlayHintRefresh = Bool |? (InlayHintOptions |? InlayHintRegistrationOptions)

ServerCapability 'Method_TextDocumentDiagnostic = DiagnosticOptions |? DiagnosticRegistrationOptions

ServerCapability 'Method_WorkspaceDiagnostic = DiagnosticOptions |? DiagnosticRegistrationOptions

ServerCapability 'Method_WorkspaceDiagnosticRefresh = DiagnosticOptions |? DiagnosticRegistrationOptions

ServerCapability 'Method_WorkspaceWillCreateFiles = FileOperationOptions

ServerCapability 'Method_WorkspaceWillRenameFiles = FileOperationOptions

ServerCapability 'Method_WorkspaceWillDeleteFiles = FileOperationOptions

ServerCapability 'Method_WorkspaceDidCreateFiles = FileOperationOptions

ServerCapability 'Method_WorkspaceDidRenameFiles = FileOperationOptions

ServerCapability 'Method_WorkspaceDidDeleteFiles = FileOperationOptions

ServerCapability 'Method_TextDocumentDidOpen = DocumentSyncCaps

ServerCapability 'Method_TextDocumentDidChange = DocumentSyncCaps

ServerCapability 'Method_TextDocumentDidClose = DocumentSyncCaps

ServerCapability 'Method_TextDocumentDidSave = DocumentSyncCaps

ServerCapability 'Method_TextDocumentWillSave = DocumentSyncCaps

ServerCapability 'Method_TextDocumentWillSaveWaitUntil = DocumentSyncCaps

ServerCapability 'Method_NotebookDocumentDidOpen = NotebookDocumentSyncCaps

ServerCapability 'Method_NotebookDocumentDidChange = NotebookDocumentSyncCaps

ServerCapability 'Method_NotebookDocumentDidSave = NotebookDocumentSyncCaps

ServerCapability 'Method_NotebookDocumentDidClose = NotebookDocumentSyncCaps

ServerCapability 'Method_WorkspaceDidChangeConfiguration = Void

ServerCapability 'Method_WorkspaceDidChangeWatchedFiles = Void

ServerCapability 'Method_WorkspaceDidChangeWorkspaceFolders = Void

ServerCapability 'Method_TextDocumentPublishDiagnostics = Void

ServerCapability 'Method_WorkspaceConfiguration = Void

ServerCapability 'Method_WindowWorkDoneProgressCreate = Void

ServerCapability 'Method_WindowWorkDoneProgressCancel = Void

ServerCapability ('Method_Progress :: Method f 'Notification) = Void

ServerCapability 'Method_WindowShowDocument = Void

ServerCapability 'Method_WindowShowMessageRequest = Void

ServerCapability 'Method_WindowShowMessage = Void

ServerCapability 'Method_WindowLogMessage = Void

ServerCapability 'Method_ClientRegisterCapability = Void

ServerCapability 'Method_ClientUnregisterCapability = Void

ServerCapability 'Method_Initialize = Void

ServerCapability 'Method_Initialized = Void

ServerCapability 'Method_Shutdown = Void

ServerCapability 'Method_Exit = Void

ServerCapability 'Method_WorkspaceApplyEdit = Void

ServerCapability 'Method_TelemetryEvent = Void

ServerCapability 'Method_SetTrace = Void

ServerCapability 'Method_LogTrace = Void

ServerCapability ('Method_CancelRequest :: Method f 'Notification) = Void

ServerCapability ('Method_CustomMethod s :: Method f t) = Void

serverCapability :: forall {f :: MessageDirection} {t :: MessageKind} (m :: Method f t). SMethod m -> Lens' ServerCapabilities (Maybe (ServerCapability m))

A lens which focusses on the (possibly absent) server capability associated with a method.