Language.LSP.Protocol.Message

Messages

LSP protocol message types and metadata

data NotificationMessage

Notification message type as defined in the spec.

Constructors

NotificationMessage

Fields

_jsonrpc :: Text

_method :: Text

_params :: Maybe Value

data RequestMessage

Request message type as defined in the spec.

Constructors

RequestMessage

Fields

_jsonrpc :: Text

_id :: Int32 |? Text

_method :: Text

_params :: Maybe Value

data ResponseError

Response error type as defined in the spec.

Constructors

ResponseError

Fields

_code :: LSPErrorCodes |? ErrorCodes

_message :: Text

_xdata :: Maybe Value

data ResponseMessage

Response message type as defined in the spec.

Constructors

ResponseMessage

Fields

_jsonrpc :: Text

_id :: Int32 |? (Text |? Null)

_result :: Maybe Value

_error :: Maybe ResponseError

data TNotificationMessage (m :: Method f Notification)

Typed notification message, containing the correct parameter payload.

Constructors

TNotificationMessage

Fields

_jsonrpc :: Text

_method :: SMethod m

_params :: MessageParams m

data TRequestMessage (m :: Method f Request)

Typed request message, containing the correct parameter payload.

Constructors

TRequestMessage

Fields

_jsonrpc :: Text

_id :: LspId m

_method :: SMethod m

_params :: MessageParams m

data TResponseError (m :: Method f Request)

Constructors

TResponseError

Fields

_code :: LSPErrorCodes |? ErrorCodes

_message :: Text

_xdata :: Maybe (ErrorData m)

toUntypedResponseError :: ToJSON (ErrorData m) => TResponseError m -> ResponseError

data TResponseMessage (m :: Method f Request)

A typed response message with a correct result payload.

Constructors

TResponseMessage

Fields

_jsonrpc :: Text

_id :: Maybe (LspId m)

_result :: Either ResponseError (MessageResult m)

data TCustomMessage s f t where

A typed custom message. A special data type is needed to distinguish between notifications and requests, since a CustomMethod can be both!

Constructors

ReqMess :: TRequestMessage (Method_CustomMethod s :: Method f Request) -> TCustomMessage s f Request

NotMess :: TNotificationMessage (Method_CustomMethod s :: Method f Notification) -> TCustomMessage s f Notification

type family TMessage m where ...

Map a method to the Request/Notification type with the correct payload.

Equations

TMessage (Method_CustomMethod s :: Method f t) = TCustomMessage s f t

TMessage (m :: Method f Request) = TRequestMessage m

TMessage (m :: Method f Notification) = TNotificationMessage m

type TClientMessage (m :: Method ClientToServer t) = TMessage m

type TServerMessage (m :: Method ServerToClient t) = TMessage m

addNullField :: String -> Value -> Value

Replace a missing field in an object with a null field, to simplify parsing This is a hack to allow other types than Maybe to work like Maybe in allowing the field to be missing. See also this issue: https://github.com/haskell/aeson/issues/646

data LspId (m :: Method f Request)

Id used for a request, Can be either a String or an Int

Constructors

IdInt !Int32

IdString !Text

data SomeLspId where

Constructors

SomeLspId :: !(LspId m) -> SomeLspId

data MessageDirection

Which direction messages are sent in.

Constructors

ServerToClient

ClientToServer

data MessageKind

What kind of message is sent.

Constructors

Notification

Request

data SMessageDirection (f :: MessageDirection) where

Singleton type for MessageDirection.

Constructors

SClientToServer :: SMessageDirection ClientToServer

SServerToClient :: SMessageDirection ServerToClient

SBothDirections :: SMessageDirection f

data SMessageKind (f :: MessageKind) where

Singleton type for MessageKind.

Constructors

SNotification :: SMessageKind Notification

SRequest :: SMessageKind Request

SBothTypes :: SMessageKind f

Parsing LSP messages

data FromServerMessage' a where

Constructors

FromServerMess :: forall t (m :: Method ServerToClient t) a. SMethod m -> TMessage m -> FromServerMessage' a

FromServerRsp :: forall (m :: Method ClientToServer Request) a. a m -> TResponseMessage m -> FromServerMessage' a

type FromServerMessage = FromServerMessage' SMethod

fromServerNot :: forall (m :: Method ServerToClient Notification). TMessage m ~ TNotificationMessage m => TNotificationMessage m -> FromServerMessage

fromServerReq :: forall (m :: Method ServerToClient Request). TMessage m ~ TRequestMessage m => TRequestMessage m -> FromServerMessage

data FromClientMessage' a where

Constructors

FromClientMess :: forall t (m :: Method ClientToServer t) a. SMethod m -> TMessage m -> FromClientMessage' a

FromClientRsp :: forall (m :: Method ServerToClient Request) a. a m -> TResponseMessage m -> FromClientMessage' a

type FromClientMessage = FromClientMessage' SMethod

fromClientNot :: forall (m :: Method ClientToServer Notification). TMessage m ~ TNotificationMessage m => TNotificationMessage m -> FromClientMessage

fromClientReq :: forall (m :: Method ClientToServer Request). TMessage m ~ TRequestMessage m => TRequestMessage m -> FromClientMessage

type LookupFunc f a = forall (m :: Method f Request). LspId m -> Maybe (SMethod m, a m)

parseServerMessage :: LookupFunc ClientToServer a -> Value -> Parser (FromServerMessage' a)

parseClientMessage :: LookupFunc ServerToClient a -> Value -> Parser (FromClientMessage' a)

clientResponseJSON :: SClientMethod m -> (HasJSON (TResponseMessage m) => x) -> x

serverResponseJSON :: SServerMethod m -> (HasJSON (TResponseMessage m) => x) -> x

clientMethodJSON :: SClientMethod m -> (ToJSON (TClientMessage m) => x) -> x

serverMethodJSON :: SServerMethod m -> (ToJSON (TServerMessage m) => x) -> x

type HasJSON a = (ToJSON a, FromJSON a, Eq a)

data ClientNotOrReq m where

Constructors

IsClientNot :: (HasJSON (TClientMessage m), TMessage m ~ TNotificationMessage m) => ClientNotOrReq (m :: Method ClientToServer Notification)

IsClientReq :: forall (m :: Method ClientToServer Request). (HasJSON (TClientMessage m), HasJSON (TResponseMessage m), TMessage m ~ TRequestMessage m) => ClientNotOrReq m

IsClientEither :: ClientNotOrReq (Method_CustomMethod s)

data ServerNotOrReq m where

Constructors

IsServerNot :: (HasJSON (TServerMessage m), TMessage m ~ TNotificationMessage m) => ServerNotOrReq (m :: Method ServerToClient Notification)

IsServerReq :: forall (m :: Method ServerToClient Request). (HasJSON (TServerMessage m), HasJSON (TResponseMessage m), TMessage m ~ TRequestMessage m) => ServerNotOrReq m

IsServerEither :: ServerNotOrReq (Method_CustomMethod s)

splitClientMethod :: SClientMethod m -> ClientNotOrReq m

splitServerMethod :: SServerMethod m -> ServerNotOrReq m

data CustomEq m1 m2 where

Given a witness that two custom methods are of the same type, produce a witness that the methods are the same

Constructors

CustomEq

Fields

:: (m1 ~ (Method_CustomMethod s :: Method f t1), m2 ~ (Method_CustomMethod s :: Method f t2))

=> { runCustomEq :: t1 ~ t2 => m1 :~~: m2

} -> CustomEq m1 m2

runEq :: t1 ~ t2 => (SMethod m1 -> SMethod m2 -> Maybe (Either (CustomEq m1 m2) (m1 :~~: m2))) -> SMethod (m1 :: Method f t1) -> SMethod (m2 :: Method f t2) -> Maybe (m1 :~~: m2)

mEqServer :: SServerMethod m1 -> SServerMethod m2 -> Maybe (Either (CustomEq m1 m2) (m1 :~~: m2))

Heterogeneous equality on singleton server methods

mEqClient :: SClientMethod m1 -> SClientMethod m2 -> Maybe (Either (CustomEq m1 m2) (m1 :~~: m2))

Heterogeneous equality on singleton client methods

Methods

Main LSP method types and functions

data Method f t where

A type representing a LSP method (or class of methods), intended to be used mostly at the type level.

Constructors

Method_TextDocumentImplementation :: Method ClientToServer Request

Method_TextDocumentTypeDefinition :: Method ClientToServer Request

Method_WorkspaceWorkspaceFolders :: Method ServerToClient Request

Method_WorkspaceConfiguration :: Method ServerToClient Request

Method_TextDocumentDocumentColor :: Method ClientToServer Request

Method_TextDocumentColorPresentation :: Method ClientToServer Request

Method_TextDocumentFoldingRange :: Method ClientToServer Request

Method_TextDocumentDeclaration :: Method ClientToServer Request

Method_TextDocumentSelectionRange :: Method ClientToServer Request

Method_WindowWorkDoneProgressCreate :: Method ServerToClient Request

Method_TextDocumentPrepareCallHierarchy :: Method ClientToServer Request

Method_CallHierarchyIncomingCalls :: Method ClientToServer Request

Method_CallHierarchyOutgoingCalls :: Method ClientToServer Request

Method_TextDocumentSemanticTokensFull :: Method ClientToServer Request

Method_TextDocumentSemanticTokensFullDelta :: Method ClientToServer Request

Method_TextDocumentSemanticTokensRange :: Method ClientToServer Request

Method_WorkspaceSemanticTokensRefresh :: Method ServerToClient Request

Method_WindowShowDocument :: Method ServerToClient Request

Method_TextDocumentLinkedEditingRange :: Method ClientToServer Request

Method_WorkspaceWillCreateFiles :: Method ClientToServer Request

Method_WorkspaceWillRenameFiles :: Method ClientToServer Request

Method_WorkspaceWillDeleteFiles :: Method ClientToServer Request

Method_TextDocumentMoniker :: Method ClientToServer Request

Method_TextDocumentPrepareTypeHierarchy :: Method ClientToServer Request

Method_TypeHierarchySupertypes :: Method ClientToServer Request

Method_TypeHierarchySubtypes :: Method ClientToServer Request

Method_TextDocumentInlineValue :: Method ClientToServer Request

Method_WorkspaceInlineValueRefresh :: Method ServerToClient Request

Method_TextDocumentInlayHint :: Method ClientToServer Request

Method_InlayHintResolve :: Method ClientToServer Request

Method_WorkspaceInlayHintRefresh :: Method ServerToClient Request

Method_TextDocumentDiagnostic :: Method ClientToServer Request

Method_WorkspaceDiagnostic :: Method ClientToServer Request

Method_WorkspaceDiagnosticRefresh :: Method ServerToClient Request

Method_ClientRegisterCapability :: Method ServerToClient Request

Method_ClientUnregisterCapability :: Method ServerToClient Request

Method_Initialize :: Method ClientToServer Request

Method_Shutdown :: Method ClientToServer Request

Method_WindowShowMessageRequest :: Method ServerToClient Request

Method_TextDocumentWillSaveWaitUntil :: Method ClientToServer Request

Method_TextDocumentCompletion :: Method ClientToServer Request

Method_CompletionItemResolve :: Method ClientToServer Request

Method_TextDocumentHover :: Method ClientToServer Request

Method_TextDocumentSignatureHelp :: Method ClientToServer Request

Method_TextDocumentDefinition :: Method ClientToServer Request

Method_TextDocumentReferences :: Method ClientToServer Request

Method_TextDocumentDocumentHighlight :: Method ClientToServer Request

Method_TextDocumentDocumentSymbol :: Method ClientToServer Request

Method_TextDocumentCodeAction :: Method ClientToServer Request

Method_CodeActionResolve :: Method ClientToServer Request

Method_WorkspaceSymbol :: Method ClientToServer Request

Method_WorkspaceSymbolResolve :: Method ClientToServer Request

Method_TextDocumentCodeLens :: Method ClientToServer Request

Method_CodeLensResolve :: Method ClientToServer Request

Method_WorkspaceCodeLensRefresh :: Method ServerToClient Request

Method_TextDocumentDocumentLink :: Method ClientToServer Request

Method_DocumentLinkResolve :: Method ClientToServer Request

Method_TextDocumentFormatting :: Method ClientToServer Request

Method_TextDocumentRangeFormatting :: Method ClientToServer Request

Method_TextDocumentOnTypeFormatting :: Method ClientToServer Request

Method_TextDocumentRename :: Method ClientToServer Request

Method_TextDocumentPrepareRename :: Method ClientToServer Request

Method_WorkspaceExecuteCommand :: Method ClientToServer Request

Method_WorkspaceApplyEdit :: Method ServerToClient Request

Method_WorkspaceDidChangeWorkspaceFolders :: Method ClientToServer Notification

Method_WindowWorkDoneProgressCancel :: Method ClientToServer Notification

Method_WorkspaceDidCreateFiles :: Method ClientToServer Notification

Method_WorkspaceDidRenameFiles :: Method ClientToServer Notification

Method_WorkspaceDidDeleteFiles :: Method ClientToServer Notification

Method_NotebookDocumentDidOpen :: Method ClientToServer Notification

Method_NotebookDocumentDidChange :: Method ClientToServer Notification

Method_NotebookDocumentDidSave :: Method ClientToServer Notification

Method_NotebookDocumentDidClose :: Method ClientToServer Notification

Method_Initialized :: Method ClientToServer Notification

Method_Exit :: Method ClientToServer Notification

Method_WorkspaceDidChangeConfiguration :: Method ClientToServer Notification

Method_WindowShowMessage :: Method ServerToClient Notification

Method_WindowLogMessage :: Method ServerToClient Notification

Method_TelemetryEvent :: Method ServerToClient Notification

Method_TextDocumentDidOpen :: Method ClientToServer Notification

Method_TextDocumentDidChange :: Method ClientToServer Notification

Method_TextDocumentDidClose :: Method ClientToServer Notification

Method_TextDocumentDidSave :: Method ClientToServer Notification

Method_TextDocumentWillSave :: Method ClientToServer Notification

Method_WorkspaceDidChangeWatchedFiles :: Method ClientToServer Notification

Method_TextDocumentPublishDiagnostics :: Method ServerToClient Notification

Method_SetTrace :: Method ClientToServer Notification

Method_LogTrace :: Method ServerToClient Notification

Method_CancelRequest :: Method f Notification

Method_Progress :: Method f Notification

Method_CustomMethod :: Symbol -> Method f t

type family MessageParams (m :: Method f t) where ...

Maps a LSP method to its parameter type.

Equations

MessageParams Method_TextDocumentImplementation = ImplementationParams

MessageParams Method_TextDocumentTypeDefinition = TypeDefinitionParams

MessageParams Method_WorkspaceWorkspaceFolders = Maybe Void

MessageParams Method_WorkspaceConfiguration = ConfigurationParams

MessageParams Method_TextDocumentDocumentColor = DocumentColorParams

MessageParams Method_TextDocumentColorPresentation = ColorPresentationParams

MessageParams Method_TextDocumentFoldingRange = FoldingRangeParams

MessageParams Method_TextDocumentDeclaration = DeclarationParams

MessageParams Method_TextDocumentSelectionRange = SelectionRangeParams

MessageParams Method_WindowWorkDoneProgressCreate = WorkDoneProgressCreateParams

MessageParams Method_TextDocumentPrepareCallHierarchy = CallHierarchyPrepareParams

MessageParams Method_CallHierarchyIncomingCalls = CallHierarchyIncomingCallsParams

MessageParams Method_CallHierarchyOutgoingCalls = CallHierarchyOutgoingCallsParams

MessageParams Method_TextDocumentSemanticTokensFull = SemanticTokensParams

MessageParams Method_TextDocumentSemanticTokensFullDelta = SemanticTokensDeltaParams

MessageParams Method_TextDocumentSemanticTokensRange = SemanticTokensRangeParams

MessageParams Method_WorkspaceSemanticTokensRefresh = Maybe Void

MessageParams Method_WindowShowDocument = ShowDocumentParams

MessageParams Method_TextDocumentLinkedEditingRange = LinkedEditingRangeParams

MessageParams Method_WorkspaceWillCreateFiles = CreateFilesParams

MessageParams Method_WorkspaceWillRenameFiles = RenameFilesParams

MessageParams Method_WorkspaceWillDeleteFiles = DeleteFilesParams

MessageParams Method_TextDocumentMoniker = MonikerParams

MessageParams Method_TextDocumentPrepareTypeHierarchy = TypeHierarchyPrepareParams

MessageParams Method_TypeHierarchySupertypes = TypeHierarchySupertypesParams

MessageParams Method_TypeHierarchySubtypes = TypeHierarchySubtypesParams

MessageParams Method_TextDocumentInlineValue = InlineValueParams

MessageParams Method_WorkspaceInlineValueRefresh = Maybe Void

MessageParams Method_TextDocumentInlayHint = InlayHintParams

MessageParams Method_InlayHintResolve = InlayHint

MessageParams Method_WorkspaceInlayHintRefresh = Maybe Void

MessageParams Method_TextDocumentDiagnostic = DocumentDiagnosticParams

MessageParams Method_WorkspaceDiagnostic = WorkspaceDiagnosticParams

MessageParams Method_WorkspaceDiagnosticRefresh = Maybe Void

MessageParams Method_ClientRegisterCapability = RegistrationParams

MessageParams Method_ClientUnregisterCapability = UnregistrationParams

MessageParams Method_Initialize = InitializeParams

MessageParams Method_Shutdown = Maybe Void

MessageParams Method_WindowShowMessageRequest = ShowMessageRequestParams

MessageParams Method_TextDocumentWillSaveWaitUntil = WillSaveTextDocumentParams

MessageParams Method_TextDocumentCompletion = CompletionParams

MessageParams Method_CompletionItemResolve = CompletionItem

MessageParams Method_TextDocumentHover = HoverParams

MessageParams Method_TextDocumentSignatureHelp = SignatureHelpParams

MessageParams Method_TextDocumentDefinition = DefinitionParams

MessageParams Method_TextDocumentReferences = ReferenceParams

MessageParams Method_TextDocumentDocumentHighlight = DocumentHighlightParams

MessageParams Method_TextDocumentDocumentSymbol = DocumentSymbolParams

MessageParams Method_TextDocumentCodeAction = CodeActionParams

MessageParams Method_CodeActionResolve = CodeAction

MessageParams Method_WorkspaceSymbol = WorkspaceSymbolParams

MessageParams Method_WorkspaceSymbolResolve = WorkspaceSymbol

MessageParams Method_TextDocumentCodeLens = CodeLensParams

MessageParams Method_CodeLensResolve = CodeLens

MessageParams Method_WorkspaceCodeLensRefresh = Maybe Void

MessageParams Method_TextDocumentDocumentLink = DocumentLinkParams

MessageParams Method_DocumentLinkResolve = DocumentLink

MessageParams Method_TextDocumentFormatting = DocumentFormattingParams

MessageParams Method_TextDocumentRangeFormatting = DocumentRangeFormattingParams

MessageParams Method_TextDocumentOnTypeFormatting = DocumentOnTypeFormattingParams

MessageParams Method_TextDocumentRename = RenameParams

MessageParams Method_TextDocumentPrepareRename = PrepareRenameParams

MessageParams Method_WorkspaceExecuteCommand = ExecuteCommandParams

MessageParams Method_WorkspaceApplyEdit = ApplyWorkspaceEditParams

MessageParams Method_WorkspaceDidChangeWorkspaceFolders = DidChangeWorkspaceFoldersParams

MessageParams Method_WindowWorkDoneProgressCancel = WorkDoneProgressCancelParams

MessageParams Method_WorkspaceDidCreateFiles = CreateFilesParams

MessageParams Method_WorkspaceDidRenameFiles = RenameFilesParams

MessageParams Method_WorkspaceDidDeleteFiles = DeleteFilesParams

MessageParams Method_NotebookDocumentDidOpen = DidOpenNotebookDocumentParams

MessageParams Method_NotebookDocumentDidChange = DidChangeNotebookDocumentParams

MessageParams Method_NotebookDocumentDidSave = DidSaveNotebookDocumentParams

MessageParams Method_NotebookDocumentDidClose = DidCloseNotebookDocumentParams

MessageParams Method_Initialized = InitializedParams

MessageParams Method_Exit = Maybe Void

MessageParams Method_WorkspaceDidChangeConfiguration = DidChangeConfigurationParams

MessageParams Method_WindowShowMessage = ShowMessageParams

MessageParams Method_WindowLogMessage = LogMessageParams

MessageParams Method_TelemetryEvent = Value

MessageParams Method_TextDocumentDidOpen = DidOpenTextDocumentParams

MessageParams Method_TextDocumentDidChange = DidChangeTextDocumentParams

MessageParams Method_TextDocumentDidClose = DidCloseTextDocumentParams

MessageParams Method_TextDocumentDidSave = DidSaveTextDocumentParams

MessageParams Method_TextDocumentWillSave = WillSaveTextDocumentParams

MessageParams Method_WorkspaceDidChangeWatchedFiles = DidChangeWatchedFilesParams

MessageParams Method_TextDocumentPublishDiagnostics = PublishDiagnosticsParams

MessageParams Method_SetTrace = SetTraceParams

MessageParams Method_LogTrace = LogTraceParams

MessageParams Method_CancelRequest = CancelParams

MessageParams Method_Progress = ProgressParams

MessageParams (Method_CustomMethod s) = Value

type family MessageResult (m :: Method f t) where ...

Maps a LSP method to its result type.

Equations

MessageResult Method_TextDocumentImplementation = Definition |? ([DefinitionLink] |? Null)

MessageResult Method_TextDocumentTypeDefinition = Definition |? ([DefinitionLink] |? Null)

MessageResult Method_WorkspaceWorkspaceFolders = [WorkspaceFolder] |? Null

MessageResult Method_WorkspaceConfiguration = [Value]

MessageResult Method_TextDocumentDocumentColor = [ColorInformation]

MessageResult Method_TextDocumentColorPresentation = [ColorPresentation]

MessageResult Method_TextDocumentFoldingRange = [FoldingRange] |? Null

MessageResult Method_TextDocumentDeclaration = Declaration |? ([DeclarationLink] |? Null)

MessageResult Method_TextDocumentSelectionRange = [SelectionRange] |? Null

MessageResult Method_WindowWorkDoneProgressCreate = Null

MessageResult Method_TextDocumentPrepareCallHierarchy = [CallHierarchyItem] |? Null

MessageResult Method_CallHierarchyIncomingCalls = [CallHierarchyIncomingCall] |? Null

MessageResult Method_CallHierarchyOutgoingCalls = [CallHierarchyOutgoingCall] |? Null

MessageResult Method_TextDocumentSemanticTokensFull = SemanticTokens |? Null

MessageResult Method_TextDocumentSemanticTokensFullDelta = SemanticTokens |? (SemanticTokensDelta |? Null)

MessageResult Method_TextDocumentSemanticTokensRange = SemanticTokens |? Null

MessageResult Method_WorkspaceSemanticTokensRefresh = Null

MessageResult Method_WindowShowDocument = ShowDocumentResult

MessageResult Method_TextDocumentLinkedEditingRange = LinkedEditingRanges |? Null

MessageResult Method_WorkspaceWillCreateFiles = WorkspaceEdit |? Null

MessageResult Method_WorkspaceWillRenameFiles = WorkspaceEdit |? Null

MessageResult Method_WorkspaceWillDeleteFiles = WorkspaceEdit |? Null

MessageResult Method_TextDocumentMoniker = [Moniker] |? Null

MessageResult Method_TextDocumentPrepareTypeHierarchy = [TypeHierarchyItem] |? Null

MessageResult Method_TypeHierarchySupertypes = [TypeHierarchyItem] |? Null

MessageResult Method_TypeHierarchySubtypes = [TypeHierarchyItem] |? Null

MessageResult Method_TextDocumentInlineValue = [InlineValue] |? Null

MessageResult Method_WorkspaceInlineValueRefresh = Null

MessageResult Method_TextDocumentInlayHint = [InlayHint] |? Null

MessageResult Method_InlayHintResolve = InlayHint

MessageResult Method_WorkspaceInlayHintRefresh = Null

MessageResult Method_TextDocumentDiagnostic = DocumentDiagnosticReport

MessageResult Method_WorkspaceDiagnostic = WorkspaceDiagnosticReport

MessageResult Method_WorkspaceDiagnosticRefresh = Null

MessageResult Method_ClientRegisterCapability = Null

MessageResult Method_ClientUnregisterCapability = Null

MessageResult Method_Initialize = InitializeResult

MessageResult Method_Shutdown = Null

MessageResult Method_WindowShowMessageRequest = MessageActionItem |? Null

MessageResult Method_TextDocumentWillSaveWaitUntil = [TextEdit] |? Null

MessageResult Method_TextDocumentCompletion = [CompletionItem] |? (CompletionList |? Null)

MessageResult Method_CompletionItemResolve = CompletionItem

MessageResult Method_TextDocumentHover = Hover |? Null

MessageResult Method_TextDocumentSignatureHelp = SignatureHelp |? Null

MessageResult Method_TextDocumentDefinition = Definition |? ([DefinitionLink] |? Null)

MessageResult Method_TextDocumentReferences = [Location] |? Null

MessageResult Method_TextDocumentDocumentHighlight = [DocumentHighlight] |? Null

MessageResult Method_TextDocumentDocumentSymbol = [SymbolInformation] |? ([DocumentSymbol] |? Null)

MessageResult Method_TextDocumentCodeAction = [Command |? CodeAction] |? Null

MessageResult Method_CodeActionResolve = CodeAction

MessageResult Method_WorkspaceSymbol = [SymbolInformation] |? ([WorkspaceSymbol] |? Null)

MessageResult Method_WorkspaceSymbolResolve = WorkspaceSymbol

MessageResult Method_TextDocumentCodeLens = [CodeLens] |? Null

MessageResult Method_CodeLensResolve = CodeLens

MessageResult Method_WorkspaceCodeLensRefresh = Null

MessageResult Method_TextDocumentDocumentLink = [DocumentLink] |? Null

MessageResult Method_DocumentLinkResolve = DocumentLink

MessageResult Method_TextDocumentFormatting = [TextEdit] |? Null

MessageResult Method_TextDocumentRangeFormatting = [TextEdit] |? Null

MessageResult Method_TextDocumentOnTypeFormatting = [TextEdit] |? Null

MessageResult Method_TextDocumentRename = WorkspaceEdit |? Null

MessageResult Method_TextDocumentPrepareRename = PrepareRenameResult |? Null

MessageResult Method_WorkspaceExecuteCommand = Value |? Null

MessageResult Method_WorkspaceApplyEdit = ApplyWorkspaceEditResult

MessageResult (Method_CustomMethod s) = Value

type family ErrorData (m :: Method f t) where ...

Maps a LSP method to its error data type.

Equations

ErrorData Method_TextDocumentImplementation = Maybe Void

ErrorData Method_TextDocumentTypeDefinition = Maybe Void

ErrorData Method_WorkspaceWorkspaceFolders = Maybe Void

ErrorData Method_WorkspaceConfiguration = Maybe Void

ErrorData Method_TextDocumentDocumentColor = Maybe Void

ErrorData Method_TextDocumentColorPresentation = Maybe Void

ErrorData Method_TextDocumentFoldingRange = Maybe Void

ErrorData Method_TextDocumentDeclaration = Maybe Void

ErrorData Method_TextDocumentSelectionRange = Maybe Void

ErrorData Method_WindowWorkDoneProgressCreate = Maybe Void

ErrorData Method_TextDocumentPrepareCallHierarchy = Maybe Void

ErrorData Method_CallHierarchyIncomingCalls = Maybe Void

ErrorData Method_CallHierarchyOutgoingCalls = Maybe Void

ErrorData Method_TextDocumentSemanticTokensFull = Maybe Void

ErrorData Method_TextDocumentSemanticTokensFullDelta = Maybe Void

ErrorData Method_TextDocumentSemanticTokensRange = Maybe Void

ErrorData Method_WorkspaceSemanticTokensRefresh = Maybe Void

ErrorData Method_WindowShowDocument = Maybe Void

ErrorData Method_TextDocumentLinkedEditingRange = Maybe Void

ErrorData Method_WorkspaceWillCreateFiles = Maybe Void

ErrorData Method_WorkspaceWillRenameFiles = Maybe Void

ErrorData Method_WorkspaceWillDeleteFiles = Maybe Void

ErrorData Method_TextDocumentMoniker = Maybe Void

ErrorData Method_TextDocumentPrepareTypeHierarchy = Maybe Void

ErrorData Method_TypeHierarchySupertypes = Maybe Void

ErrorData Method_TypeHierarchySubtypes = Maybe Void

ErrorData Method_TextDocumentInlineValue = Maybe Void

ErrorData Method_WorkspaceInlineValueRefresh = Maybe Void

ErrorData Method_TextDocumentInlayHint = Maybe Void

ErrorData Method_InlayHintResolve = Maybe Void

ErrorData Method_WorkspaceInlayHintRefresh = Maybe Void

ErrorData Method_TextDocumentDiagnostic = DiagnosticServerCancellationData

ErrorData Method_WorkspaceDiagnostic = DiagnosticServerCancellationData

ErrorData Method_WorkspaceDiagnosticRefresh = Maybe Void

ErrorData Method_ClientRegisterCapability = Maybe Void

ErrorData Method_ClientUnregisterCapability = Maybe Void

ErrorData Method_Initialize = InitializeError

ErrorData Method_Shutdown = Maybe Void

ErrorData Method_WindowShowMessageRequest = Maybe Void

ErrorData Method_TextDocumentWillSaveWaitUntil = Maybe Void

ErrorData Method_TextDocumentCompletion = Maybe Void

ErrorData Method_CompletionItemResolve = Maybe Void

ErrorData Method_TextDocumentHover = Maybe Void

ErrorData Method_TextDocumentSignatureHelp = Maybe Void

ErrorData Method_TextDocumentDefinition = Maybe Void

ErrorData Method_TextDocumentReferences = Maybe Void

ErrorData Method_TextDocumentDocumentHighlight = Maybe Void

ErrorData Method_TextDocumentDocumentSymbol = Maybe Void

ErrorData Method_TextDocumentCodeAction = Maybe Void

ErrorData Method_CodeActionResolve = Maybe Void

ErrorData Method_WorkspaceSymbol = Maybe Void

ErrorData Method_WorkspaceSymbolResolve = Maybe Void

ErrorData Method_TextDocumentCodeLens = Maybe Void

ErrorData Method_CodeLensResolve = Maybe Void

ErrorData Method_WorkspaceCodeLensRefresh = Maybe Void

ErrorData Method_TextDocumentDocumentLink = Maybe Void

ErrorData Method_DocumentLinkResolve = Maybe Void

ErrorData Method_TextDocumentFormatting = Maybe Void

ErrorData Method_TextDocumentRangeFormatting = Maybe Void

ErrorData Method_TextDocumentOnTypeFormatting = Maybe Void

ErrorData Method_TextDocumentRename = Maybe Void

ErrorData Method_TextDocumentPrepareRename = Maybe Void

ErrorData Method_WorkspaceExecuteCommand = Maybe Void

ErrorData Method_WorkspaceApplyEdit = Maybe Void

ErrorData (Method_CustomMethod s) = Value

type family RegistrationOptions (m :: Method f t) where ...

Maps a LSP method to its registration options type.

Equations

RegistrationOptions Method_TextDocumentImplementation = ImplementationRegistrationOptions

RegistrationOptions Method_TextDocumentTypeDefinition = TypeDefinitionRegistrationOptions

RegistrationOptions Method_WorkspaceWorkspaceFolders = Maybe Void

RegistrationOptions Method_WorkspaceConfiguration = Maybe Void

RegistrationOptions Method_TextDocumentDocumentColor = DocumentColorRegistrationOptions

RegistrationOptions Method_TextDocumentColorPresentation = Rec (("workDoneProgress" .== Maybe Bool) .+ (("documentSelector" .== (DocumentSelector |? Null)) .+ Empty))

RegistrationOptions Method_TextDocumentFoldingRange = FoldingRangeRegistrationOptions

RegistrationOptions Method_TextDocumentDeclaration = DeclarationRegistrationOptions

RegistrationOptions Method_TextDocumentSelectionRange = SelectionRangeRegistrationOptions

RegistrationOptions Method_WindowWorkDoneProgressCreate = Maybe Void

RegistrationOptions Method_TextDocumentPrepareCallHierarchy = CallHierarchyRegistrationOptions

RegistrationOptions Method_CallHierarchyIncomingCalls = Maybe Void

RegistrationOptions Method_CallHierarchyOutgoingCalls = Maybe Void

RegistrationOptions Method_TextDocumentSemanticTokensFull = SemanticTokensRegistrationOptions

RegistrationOptions Method_TextDocumentSemanticTokensFullDelta = SemanticTokensRegistrationOptions

RegistrationOptions Method_TextDocumentSemanticTokensRange = Maybe Void

RegistrationOptions Method_WorkspaceSemanticTokensRefresh = Maybe Void

RegistrationOptions Method_WindowShowDocument = Maybe Void

RegistrationOptions Method_TextDocumentLinkedEditingRange = LinkedEditingRangeRegistrationOptions

RegistrationOptions Method_WorkspaceWillCreateFiles = FileOperationRegistrationOptions

RegistrationOptions Method_WorkspaceWillRenameFiles = FileOperationRegistrationOptions

RegistrationOptions Method_WorkspaceWillDeleteFiles = FileOperationRegistrationOptions

RegistrationOptions Method_TextDocumentMoniker = MonikerRegistrationOptions

RegistrationOptions Method_TextDocumentPrepareTypeHierarchy = TypeHierarchyRegistrationOptions

RegistrationOptions Method_TypeHierarchySupertypes = Maybe Void

RegistrationOptions Method_TypeHierarchySubtypes = Maybe Void

RegistrationOptions Method_TextDocumentInlineValue = InlineValueRegistrationOptions

RegistrationOptions Method_WorkspaceInlineValueRefresh = Maybe Void

RegistrationOptions Method_TextDocumentInlayHint = InlayHintRegistrationOptions

RegistrationOptions Method_InlayHintResolve = Maybe Void

RegistrationOptions Method_WorkspaceInlayHintRefresh = Maybe Void

RegistrationOptions Method_TextDocumentDiagnostic = DiagnosticRegistrationOptions

RegistrationOptions Method_WorkspaceDiagnostic = Maybe Void

RegistrationOptions Method_WorkspaceDiagnosticRefresh = Maybe Void

RegistrationOptions Method_ClientRegisterCapability = Maybe Void

RegistrationOptions Method_ClientUnregisterCapability = Maybe Void

RegistrationOptions Method_Initialize = Maybe Void

RegistrationOptions Method_Shutdown = Maybe Void

RegistrationOptions Method_WindowShowMessageRequest = Maybe Void

RegistrationOptions Method_TextDocumentWillSaveWaitUntil = TextDocumentRegistrationOptions

RegistrationOptions Method_TextDocumentCompletion = CompletionRegistrationOptions

RegistrationOptions Method_CompletionItemResolve = Maybe Void

RegistrationOptions Method_TextDocumentHover = HoverRegistrationOptions

RegistrationOptions Method_TextDocumentSignatureHelp = SignatureHelpRegistrationOptions

RegistrationOptions Method_TextDocumentDefinition = DefinitionRegistrationOptions

RegistrationOptions Method_TextDocumentReferences = ReferenceRegistrationOptions

RegistrationOptions Method_TextDocumentDocumentHighlight = DocumentHighlightRegistrationOptions

RegistrationOptions Method_TextDocumentDocumentSymbol = DocumentSymbolRegistrationOptions

RegistrationOptions Method_TextDocumentCodeAction = CodeActionRegistrationOptions

RegistrationOptions Method_CodeActionResolve = Maybe Void

RegistrationOptions Method_WorkspaceSymbol = WorkspaceSymbolRegistrationOptions

RegistrationOptions Method_WorkspaceSymbolResolve = Maybe Void

RegistrationOptions Method_TextDocumentCodeLens = CodeLensRegistrationOptions

RegistrationOptions Method_CodeLensResolve = Maybe Void

RegistrationOptions Method_WorkspaceCodeLensRefresh = Maybe Void

RegistrationOptions Method_TextDocumentDocumentLink = DocumentLinkRegistrationOptions

RegistrationOptions Method_DocumentLinkResolve = Maybe Void

RegistrationOptions Method_TextDocumentFormatting = DocumentFormattingRegistrationOptions

RegistrationOptions Method_TextDocumentRangeFormatting = DocumentRangeFormattingRegistrationOptions

RegistrationOptions Method_TextDocumentOnTypeFormatting = DocumentOnTypeFormattingRegistrationOptions

RegistrationOptions Method_TextDocumentRename = RenameRegistrationOptions

RegistrationOptions Method_TextDocumentPrepareRename = Maybe Void

RegistrationOptions Method_WorkspaceExecuteCommand = ExecuteCommandRegistrationOptions

RegistrationOptions Method_WorkspaceApplyEdit = Maybe Void

RegistrationOptions Method_WorkspaceDidChangeWorkspaceFolders = Maybe Void

RegistrationOptions Method_WindowWorkDoneProgressCancel = Maybe Void

RegistrationOptions Method_WorkspaceDidCreateFiles = FileOperationRegistrationOptions

RegistrationOptions Method_WorkspaceDidRenameFiles = FileOperationRegistrationOptions

RegistrationOptions Method_WorkspaceDidDeleteFiles = FileOperationRegistrationOptions

RegistrationOptions Method_NotebookDocumentDidOpen = Maybe Void

RegistrationOptions Method_NotebookDocumentDidChange = Maybe Void

RegistrationOptions Method_NotebookDocumentDidSave = Maybe Void

RegistrationOptions Method_NotebookDocumentDidClose = Maybe Void

RegistrationOptions Method_Initialized = Maybe Void

RegistrationOptions Method_Exit = Maybe Void

RegistrationOptions Method_WorkspaceDidChangeConfiguration = DidChangeConfigurationRegistrationOptions

RegistrationOptions Method_WindowShowMessage = Maybe Void

RegistrationOptions Method_WindowLogMessage = Maybe Void

RegistrationOptions Method_TelemetryEvent = Maybe Void

RegistrationOptions Method_TextDocumentDidOpen = TextDocumentRegistrationOptions

RegistrationOptions Method_TextDocumentDidChange = TextDocumentChangeRegistrationOptions

RegistrationOptions Method_TextDocumentDidClose = TextDocumentRegistrationOptions

RegistrationOptions Method_TextDocumentDidSave = TextDocumentSaveRegistrationOptions

RegistrationOptions Method_TextDocumentWillSave = TextDocumentRegistrationOptions

RegistrationOptions Method_WorkspaceDidChangeWatchedFiles = DidChangeWatchedFilesRegistrationOptions

RegistrationOptions Method_TextDocumentPublishDiagnostics = Maybe Void

RegistrationOptions Method_SetTrace = Maybe Void

RegistrationOptions Method_LogTrace = Maybe Void

RegistrationOptions Method_CancelRequest = Maybe Void

RegistrationOptions Method_Progress = Maybe Void

RegistrationOptions (Method_CustomMethod s) = Void

data SMethod m where

A singleton type for Method.

Constructors

SMethod_TextDocumentImplementation :: SMethod Method_TextDocumentImplementation

SMethod_TextDocumentTypeDefinition :: SMethod Method_TextDocumentTypeDefinition

SMethod_WorkspaceWorkspaceFolders :: SMethod Method_WorkspaceWorkspaceFolders

SMethod_WorkspaceConfiguration :: SMethod Method_WorkspaceConfiguration

SMethod_TextDocumentDocumentColor :: SMethod Method_TextDocumentDocumentColor

SMethod_TextDocumentColorPresentation :: SMethod Method_TextDocumentColorPresentation

SMethod_TextDocumentFoldingRange :: SMethod Method_TextDocumentFoldingRange

SMethod_TextDocumentDeclaration :: SMethod Method_TextDocumentDeclaration

SMethod_TextDocumentSelectionRange :: SMethod Method_TextDocumentSelectionRange

SMethod_WindowWorkDoneProgressCreate :: SMethod Method_WindowWorkDoneProgressCreate

SMethod_TextDocumentPrepareCallHierarchy :: SMethod Method_TextDocumentPrepareCallHierarchy

SMethod_CallHierarchyIncomingCalls :: SMethod Method_CallHierarchyIncomingCalls

SMethod_CallHierarchyOutgoingCalls :: SMethod Method_CallHierarchyOutgoingCalls

SMethod_TextDocumentSemanticTokensFull :: SMethod Method_TextDocumentSemanticTokensFull

SMethod_TextDocumentSemanticTokensFullDelta :: SMethod Method_TextDocumentSemanticTokensFullDelta

SMethod_TextDocumentSemanticTokensRange :: SMethod Method_TextDocumentSemanticTokensRange

SMethod_WorkspaceSemanticTokensRefresh :: SMethod Method_WorkspaceSemanticTokensRefresh

SMethod_WindowShowDocument :: SMethod Method_WindowShowDocument

SMethod_TextDocumentLinkedEditingRange :: SMethod Method_TextDocumentLinkedEditingRange

SMethod_WorkspaceWillCreateFiles :: SMethod Method_WorkspaceWillCreateFiles

SMethod_WorkspaceWillRenameFiles :: SMethod Method_WorkspaceWillRenameFiles

SMethod_WorkspaceWillDeleteFiles :: SMethod Method_WorkspaceWillDeleteFiles

SMethod_TextDocumentMoniker :: SMethod Method_TextDocumentMoniker

SMethod_TextDocumentPrepareTypeHierarchy :: SMethod Method_TextDocumentPrepareTypeHierarchy

SMethod_TypeHierarchySupertypes :: SMethod Method_TypeHierarchySupertypes

SMethod_TypeHierarchySubtypes :: SMethod Method_TypeHierarchySubtypes

SMethod_TextDocumentInlineValue :: SMethod Method_TextDocumentInlineValue

SMethod_WorkspaceInlineValueRefresh :: SMethod Method_WorkspaceInlineValueRefresh

SMethod_TextDocumentInlayHint :: SMethod Method_TextDocumentInlayHint

SMethod_InlayHintResolve :: SMethod Method_InlayHintResolve

SMethod_WorkspaceInlayHintRefresh :: SMethod Method_WorkspaceInlayHintRefresh

SMethod_TextDocumentDiagnostic :: SMethod Method_TextDocumentDiagnostic

SMethod_WorkspaceDiagnostic :: SMethod Method_WorkspaceDiagnostic

SMethod_WorkspaceDiagnosticRefresh :: SMethod Method_WorkspaceDiagnosticRefresh

SMethod_ClientRegisterCapability :: SMethod Method_ClientRegisterCapability

SMethod_ClientUnregisterCapability :: SMethod Method_ClientUnregisterCapability

SMethod_Initialize :: SMethod Method_Initialize

SMethod_Shutdown :: SMethod Method_Shutdown

SMethod_WindowShowMessageRequest :: SMethod Method_WindowShowMessageRequest

SMethod_TextDocumentWillSaveWaitUntil :: SMethod Method_TextDocumentWillSaveWaitUntil

SMethod_TextDocumentCompletion :: SMethod Method_TextDocumentCompletion

SMethod_CompletionItemResolve :: SMethod Method_CompletionItemResolve

SMethod_TextDocumentHover :: SMethod Method_TextDocumentHover

SMethod_TextDocumentSignatureHelp :: SMethod Method_TextDocumentSignatureHelp

SMethod_TextDocumentDefinition :: SMethod Method_TextDocumentDefinition

SMethod_TextDocumentReferences :: SMethod Method_TextDocumentReferences

SMethod_TextDocumentDocumentHighlight :: SMethod Method_TextDocumentDocumentHighlight

SMethod_TextDocumentDocumentSymbol :: SMethod Method_TextDocumentDocumentSymbol

SMethod_TextDocumentCodeAction :: SMethod Method_TextDocumentCodeAction

SMethod_CodeActionResolve :: SMethod Method_CodeActionResolve

SMethod_WorkspaceSymbol :: SMethod Method_WorkspaceSymbol

SMethod_WorkspaceSymbolResolve :: SMethod Method_WorkspaceSymbolResolve

SMethod_TextDocumentCodeLens :: SMethod Method_TextDocumentCodeLens

SMethod_CodeLensResolve :: SMethod Method_CodeLensResolve

SMethod_WorkspaceCodeLensRefresh :: SMethod Method_WorkspaceCodeLensRefresh

SMethod_TextDocumentDocumentLink :: SMethod Method_TextDocumentDocumentLink

SMethod_DocumentLinkResolve :: SMethod Method_DocumentLinkResolve

SMethod_TextDocumentFormatting :: SMethod Method_TextDocumentFormatting

SMethod_TextDocumentRangeFormatting :: SMethod Method_TextDocumentRangeFormatting

SMethod_TextDocumentOnTypeFormatting :: SMethod Method_TextDocumentOnTypeFormatting

SMethod_TextDocumentRename :: SMethod Method_TextDocumentRename

SMethod_TextDocumentPrepareRename :: SMethod Method_TextDocumentPrepareRename

SMethod_WorkspaceExecuteCommand :: SMethod Method_WorkspaceExecuteCommand

SMethod_WorkspaceApplyEdit :: SMethod Method_WorkspaceApplyEdit

SMethod_WorkspaceDidChangeWorkspaceFolders :: SMethod Method_WorkspaceDidChangeWorkspaceFolders

SMethod_WindowWorkDoneProgressCancel :: SMethod Method_WindowWorkDoneProgressCancel

SMethod_WorkspaceDidCreateFiles :: SMethod Method_WorkspaceDidCreateFiles

SMethod_WorkspaceDidRenameFiles :: SMethod Method_WorkspaceDidRenameFiles

SMethod_WorkspaceDidDeleteFiles :: SMethod Method_WorkspaceDidDeleteFiles

SMethod_NotebookDocumentDidOpen :: SMethod Method_NotebookDocumentDidOpen

SMethod_NotebookDocumentDidChange :: SMethod Method_NotebookDocumentDidChange

SMethod_NotebookDocumentDidSave :: SMethod Method_NotebookDocumentDidSave

SMethod_NotebookDocumentDidClose :: SMethod Method_NotebookDocumentDidClose

SMethod_Initialized :: SMethod Method_Initialized

SMethod_Exit :: SMethod Method_Exit

SMethod_WorkspaceDidChangeConfiguration :: SMethod Method_WorkspaceDidChangeConfiguration

SMethod_WindowShowMessage :: SMethod Method_WindowShowMessage

SMethod_WindowLogMessage :: SMethod Method_WindowLogMessage

SMethod_TelemetryEvent :: SMethod Method_TelemetryEvent

SMethod_TextDocumentDidOpen :: SMethod Method_TextDocumentDidOpen

SMethod_TextDocumentDidChange :: SMethod Method_TextDocumentDidChange

SMethod_TextDocumentDidClose :: SMethod Method_TextDocumentDidClose

SMethod_TextDocumentDidSave :: SMethod Method_TextDocumentDidSave

SMethod_TextDocumentWillSave :: SMethod Method_TextDocumentWillSave

SMethod_WorkspaceDidChangeWatchedFiles :: SMethod Method_WorkspaceDidChangeWatchedFiles

SMethod_TextDocumentPublishDiagnostics :: SMethod Method_TextDocumentPublishDiagnostics

SMethod_SetTrace :: SMethod Method_SetTrace

SMethod_LogTrace :: SMethod Method_LogTrace

SMethod_CancelRequest :: SMethod Method_CancelRequest

SMethod_Progress :: SMethod Method_Progress

SMethod_CustomMethod :: forall s. KnownSymbol s => Proxy s -> SMethod (Method_CustomMethod s)

data SomeMethod where

A method which isn't statically known.

Constructors

SomeMethod :: forall m. SMethod m -> SomeMethod

someMethodToMethodString :: SomeMethod -> String

Turn a SomeMethod into its LSP method string.

methodStringToSomeMethod :: String -> SomeMethod

Turn a LSP method string into a SomeMethod.

messageDirection :: forall f t (m :: Method f t). SMethod m -> SMessageDirection f

Get a singleton witness for the message direction of a SMethod.

messageKind :: forall f t (m :: Method f t). SMethod m -> SMessageKind t

Get a singleton witness for the message kind of a SMethod.

Helpers for working with methods

isOptionalMethod :: SomeMethod -> Bool

Is this an "optional" method which servers and clients are allowed to ignore?

type SClientMethod (m :: Method ClientToServer t) = SMethod m

type SServerMethod (m :: Method ServerToClient t) = SMethod m

data SomeClientMethod

Constructors

forall t (m :: Method ClientToServer t). SomeClientMethod (SMethod m)

data SomeServerMethod

Constructors

forall t (m :: Method ServerToClient t). SomeServerMethod (SMethod m)

someClientMethod :: SMethod m -> Maybe SomeClientMethod

someServerMethod :: SMethod m -> Maybe SomeServerMethod

LSP registrations

data TRegistration (m :: Method ClientToServer t)

Typed registration type, with correct options.

Constructors

TRegistration

Fields

_id :: Text

The id used to register the request. The id can be used to deregister the request again.

_method :: SClientMethod m

The method / capability to register for.

_registerOptions :: !(Maybe (RegistrationOptions m))

Options necessary for the registration. Make this strict to aid the pattern matching exhaustiveness checker

regHelper :: forall m x. SMethod m -> (Show (RegistrationOptions m) => ToJSON (RegistrationOptions m) => FromJSON (RegistrationOptions m) => x) -> x

data SomeRegistration

Constructors

forall t (m :: Method ClientToServer t). SomeRegistration (TRegistration m)

toUntypedRegistration :: TRegistration m -> Registration

toSomeRegistration :: Registration -> Maybe SomeRegistration

data TUnregistration (m :: Method ClientToServer t)

Typed unregistration type.

Constructors

TUnregistration

Fields

_id :: Text

The id used to unregister the request or notification. Usually an id provided during the register request.

_method :: SMethod m

The method / capability to unregister for.

data SomeUnregistration

Constructors

forall t (m :: Method ClientToServer t). SomeUnregistration (TUnregistration m)

toUntypedUnregistration :: TUnregistration m -> Unregistration

toSomeUnregistration :: Unregistration -> Maybe SomeUnregistration