Language.LSP.Protocol.Utils.Misc

Documentation

rdrop :: Int -> [a] -> [a]

makeSingletonFromJSON :: Name -> Name -> [Name] -> Q [Dec]

Given a wrapper and a singleton GADT, construct FromJSON instances for each constructor return type by invoking the FromJSON instance for the wrapper and unwrapping

makeRegHelper :: Name -> DecsQ

lspOptions :: Options

Standard options for use when generating JSON instances NOTE: This needs to be in a separate file because of the TH stage restriction

lspOptionsUntagged :: Options

Standard options for use when generating JSON instances for an untagged union

prettyJSON :: Value -> Doc ann

newtype ViaJSON a

Constructors

ViaJSON a

genLenses :: [Name] -> Q [Dec]

Given a list of type names, make a splice that generates the lens typeclass declarations for all of them. Defined here to avoid stage restrictions.