Language.LSP.Types.Capabilities

Documentation

data WorkspaceClientCapabilities

Constructors

WorkspaceClientCapabilities

Fields

_applyEdit :: Maybe Bool

The client supports applying batch edits to the workspace by supporting the request 'workspace/applyEdit'

_workspaceEdit :: Maybe WorkspaceEditClientCapabilities

Capabilities specific to WorkspaceEdits

_didChangeConfiguration :: Maybe DidChangeConfigurationClientCapabilities

Capabilities specific to the `workspace/didChangeConfiguration` notification.

_didChangeWatchedFiles :: Maybe DidChangeWatchedFilesClientCapabilities

Capabilities specific to the `workspace/didChangeWatchedFiles` notification.

_symbol :: Maybe WorkspaceSymbolClientCapabilities

Capabilities specific to the `workspace/symbol` request.

_executeCommand :: Maybe ExecuteCommandClientCapabilities

Capabilities specific to the `workspace/executeCommand` request.

_workspaceFolders :: Maybe Bool

The client has support for workspace folders.

_configuration :: Maybe Bool

The client supports `workspace/configuration` requests.

data TextDocumentClientCapabilities

Constructors

TextDocumentClientCapabilities

Fields

_synchronization :: Maybe TextDocumentSyncClientCapabilities

_completion :: Maybe CompletionClientCapabilities

Capabilities specific to the `textDocument/completion`

_hover :: Maybe HoverClientCapabilities

Capabilities specific to the `textDocument/hover`

_signatureHelp :: Maybe SignatureHelpClientCapabilities

Capabilities specific to the `textDocument/signatureHelp`

_references :: Maybe ReferencesClientCapabilities

Capabilities specific to the `textDocument/references`

_documentHighlight :: Maybe DocumentHighlightClientCapabilities

Capabilities specific to the `textDocument/documentHighlight`

_documentSymbol :: Maybe DocumentSymbolClientCapabilities

Capabilities specific to the `textDocument/documentSymbol`

_formatting :: Maybe DocumentFormattingClientCapabilities

Capabilities specific to the `textDocument/formatting`

_rangeFormatting :: Maybe DocumentRangeFormattingClientCapabilities

Capabilities specific to the `textDocument/rangeFormatting`

_onTypeFormatting :: Maybe DocumentOnTypeFormattingClientCapabilities

Capabilities specific to the `textDocument/onTypeFormatting`

_declaration :: Maybe DeclarationClientCapabilities

Capabilities specific to the `textDocument/declaration` request.

Since LSP 3.14.0

_definition :: Maybe DefinitionClientCapabilities

Capabilities specific to the `textDocument/definition`

_typeDefinition :: Maybe TypeDefinitionClientCapabilities

Capabilities specific to the `textDocument/typeDefinition`

_implementation :: Maybe ImplementationClientCapabilities

Capabilities specific to the `textDocument/implementation`

_codeAction :: Maybe CodeActionClientCapabilities

Capabilities specific to the `textDocument/codeAction`

_codeLens :: Maybe CodeLensClientCapabilities

Capabilities specific to the `textDocument/codeLens`

_documentLink :: Maybe DocumentLinkClientCapabilities

Capabilities specific to the `textDocument/documentLink`

_colorProvider :: Maybe DocumentColorClientCapabilities

Capabilities specific to the `textDocument/documentColor` and the `textDocument/colorPresentation` request

_rename :: Maybe RenameClientCapabilities

Capabilities specific to the `textDocument/rename`

_publishDiagnostics :: Maybe PublishDiagnosticsClientCapabilities

Capabilities specific to `textDocument/publishDiagnostics`

_foldingRange :: Maybe FoldingRangeClientCapabilities

Capabilities specific to the `textDocument/foldingRange` request. Since LSP 3.10.

Since: 0.7.0.0

_selectionRange :: Maybe SelectionRangeClientCapabilities

Capabilities specific to the `textDocument/selectionRange` request. Since LSP 3.15.0

data WindowClientCapabilities

Window specific client capabilities.

Constructors

WindowClientCapabilities

Fields

_workDoneProgress :: Maybe Bool

Whether client supports handling progress notifications.

data ClientCapabilities

Constructors

ClientCapabilities

Fields

_workspace :: Maybe WorkspaceClientCapabilities

_textDocument :: Maybe TextDocumentClientCapabilities

_window :: Maybe WindowClientCapabilities

Capabilities specific to `window/progress` requests. Experimental.

Since: 0.10.0.0

_experimental :: Maybe Object

data WorkspaceFoldersServerCapabilities

Constructors

WorkspaceFoldersServerCapabilities

Fields

_supported :: Maybe Bool

The server has support for workspace folders

_changeNotifications :: Maybe (Text |? Bool)

Whether the server wants to receive workspace folder change notifications. If a strings is provided the string is treated as a ID under which the notification is registered on the client side. The ID can be used to unregister for these events using the `client/unregisterCapability` request.

data WorkspaceServerCapabilities

Constructors

WorkspaceServerCapabilities

Fields

_workspaceFolders :: Maybe WorkspaceFoldersServerCapabilities

The server supports workspace folder. Since LSP 3.6

Since: 0.7.0.0

data ServerCapabilities

Constructors

ServerCapabilities

Fields

_textDocumentSync :: Maybe (TextDocumentSyncOptions |? TextDocumentSyncKind)

Defines how text documents are synced. Is either a detailed structure defining each notification or for backwards compatibility the TextDocumentSyncKind number. If omitted it defaults to TdSyncNone.

_hoverProvider :: Maybe (Bool |? HoverOptions)

The server provides hover support.

_completionProvider :: Maybe CompletionOptions

The server provides completion support.

_signatureHelpProvider :: Maybe SignatureHelpOptions

The server provides signature help support.

_declarationProvider :: Maybe (Bool |? (DeclarationOptions |? DeclarationRegistrationOptions))

The server provides go to declaration support.

Since LSP 3.14.0

_definitionProvider :: Maybe (Bool |? DefinitionOptions)

The server provides goto definition support.

_typeDefinitionProvider :: Maybe (Bool |? (TypeDefinitionOptions |? TypeDefinitionRegistrationOptions))

The server provides Goto Type Definition support. Since LSP 3.6

Since: 0.7.0.0

_implementationProvider :: Maybe (Bool |? (ImplementationOptions |? ImplementationRegistrationOptions))

The server provides Goto Implementation support. Since LSP 3.6

Since: 0.7.0.0

_referencesProvider :: Maybe (Bool |? ReferenceOptions)

The server provides find references support.

_documentHighlightProvider :: Maybe (Bool |? DocumentHighlightOptions)

The server provides document highlight support.

_documentSymbolProvider :: Maybe (Bool |? DocumentSymbolOptions)

The server provides document symbol support.

_codeActionProvider :: Maybe (Bool |? CodeActionOptions)

The server provides code actions.

_codeLensProvider :: Maybe CodeLensOptions

The server provides code lens.

_documentLinkProvider :: Maybe DocumentLinkOptions

The server provides document link support.

_colorProvider :: Maybe (Bool |? (DocumentColorOptions |? DocumentColorRegistrationOptions))

The server provides color provider support. Since LSP 3.6

Since: 0.7.0.0

_documentFormattingProvider :: Maybe (Bool |? DocumentFormattingOptions)

The server provides document formatting.

_documentRangeFormattingProvider :: Maybe (Bool |? DocumentRangeFormattingOptions)

The server provides document range formatting.

_documentOnTypeFormattingProvider :: Maybe DocumentOnTypeFormattingOptions

The server provides document formatting on typing.

_renameProvider :: Maybe (Bool |? RenameOptions)

The server provides rename support.

_foldingRangeProvider :: Maybe (Bool |? (FoldingRangeOptions |? FoldingRangeRegistrationOptions))

The server provides folding provider support. Since LSP 3.10

Since: 0.7.0.0

_executeCommandProvider :: Maybe ExecuteCommandOptions

The server provides execute command support.

_selectionRangeProvider :: Maybe (Bool |? (SelectionRangeOptions |? SelectionRangeRegistrationOptions))

The server provides selection range support. Since LSP 3.15

_workspaceSymbolProvider :: Maybe Bool

The server provides workspace symbol support.

_workspace :: Maybe WorkspaceServerCapabilities

Workspace specific server capabilities

_experimental :: Maybe Value

Experimental server capabilities.

data TextEdit

Constructors

TextEdit

Fields

_range :: Range

_newText :: Text

data TextDocumentEdit

Constructors

TextDocumentEdit

Fields

_textDocument :: VersionedTextDocumentIdentifier

_edits :: List TextEdit

data CreateFileOptions

Options to create a file.

Constructors

CreateFileOptions

Fields

_overwrite :: Maybe Bool

Overwrite existing file. Overwrite wins over ignoreIfExists

_ignoreIfExists :: Maybe Bool

Ignore if exists.

data CreateFile

Create file operation

Constructors

CreateFile

Fields

_uri :: Uri

The resource to create.

_options :: Maybe CreateFileOptions

Additional options

data RenameFileOptions

Constructors

RenameFileOptions

Fields

_overwrite :: Maybe Bool

Overwrite target if existing. Overwrite wins over ignoreIfExists

_ignoreIfExists :: Maybe Bool

Ignores if target exists.

data RenameFile

Rename file operation

Constructors

RenameFile

Fields

_oldUri :: Uri

The old (existing) location.

_newUri :: Uri

The new location.

_options :: Maybe RenameFileOptions

Rename options.

data DeleteFileOptions

Constructors

DeleteFileOptions

Fields

_recursive :: Maybe Bool

Delete the content recursively if a folder is denoted.

_ignoreIfNotExists :: Maybe Bool

Ignore the operation if the file doesn't exist.

data DeleteFile

Delete file operation

Constructors

DeleteFile

Fields

_uri :: Uri

The file to delete.

_options :: Maybe DeleteFileOptions

Delete options.

type DocumentChange = TextDocumentEdit |? (CreateFile |? (RenameFile |? DeleteFile))

`TextDocumentEdit |? CreateFile |? RenameFile |? DeleteFile` is a bit mouthful, here's the synonym

type WorkspaceEditMap = HashMap Uri (List TextEdit)

data WorkspaceEdit

Constructors

WorkspaceEdit

Fields

_changes :: Maybe WorkspaceEditMap

_documentChanges :: Maybe (List DocumentChange)

data ResourceOperationKind

Constructors

ResourceOperationCreate
Supports creating new files and folders.

ResourceOperationRename
Supports renaming existing files and folders.

ResourceOperationDelete
Supports deleting existing files and folders.

data FailureHandlingKind

Constructors

FailureHandlingAbort
Applying the workspace change is simply aborted if one of the changes provided fails. All operations executed before the failing operation stay executed.

FailureHandlingTransactional
All operations are executed transactional. That means they either all succeed or no changes at all are applied to the workspace.

FailureHandlingTextOnlyTransactional
If the workspace edit contains only textual file changes they are executed transactional. If resource changes (create, rename or delete file) are part of the change the failure handling strategy is abort.

FailureHandlingUndo
The client tries to undo the operations already executed. But there is no guarantee that this is succeeding.

data WorkspaceEditClientCapabilities

Constructors

WorkspaceEditClientCapabilities

Fields

_documentChanges :: Maybe Bool

The client supports versioned document changes in WorkspaceEdits | The resource operations the client supports. Clients should at least support create, rename and delete files and folders.

_resourceOperations :: Maybe (List ResourceOperationKind)

_failureHandling :: Maybe FailureHandlingKind

The failure handling strategy of a client if applying the workspace edit fails.

data ApplyWorkspaceEditParams

Constructors

ApplyWorkspaceEditParams

Fields

_label :: Maybe Text

An optional label of the workspace edit. This label is presented in the user interface for example on an undo stack to undo the workspace edit.

_edit :: WorkspaceEdit

The edits to apply

data ApplyWorkspaceEditResponseBody

Constructors

ApplyWorkspaceEditResponseBody

Fields

_applied :: Bool

Indicates whether the edit was applied or not.

_failureReason :: Maybe Text

An optional textual description for why the edit was not applied. This may be used may be used by the server for diagnostic logging or to provide a suitable error for a request that triggered the edit.

applyTextEdit :: TextEdit -> Text -> Text

Applies a TextEdit to some Text. >>> applyTextEdit (TextEdit (Range (Position 0 1) (Position 0 2)) "i") "foo" "fio"

editTextEdit :: TextEdit -> TextEdit -> TextEdit

editTextEdit outer inner applies inner to the text inside outer.

fullCaps :: ClientCapabilities

Capabilities for full conformance to the current (v3.15) LSP specification.

data LSPVersion

A specific version of the LSP specification.

Constructors

LSPVersion Int Int
Construct a major.minor version

capsForVersion :: LSPVersion -> ClientCapabilities

Capabilities for full conformance to the LSP specification up until a version. Some important milestones:

3.12 textDocument/prepareRename request

3.11 CodeActionOptions provided by the server

3.10 hierarchical document symbols, folding ranges

3.9 completion item preselect

3.8 codeAction literals

3.7 related information in diagnostics

3.6 workspace folders, colors, goto type/implementation

3.4 extended completion item and symbol item kinds

3.0 dynamic registration