Language.LSP.VFS

Description

Handles the Language.LSP.Types.TextDocumentDidChange / Language.LSP.Types.TextDocumentDidOpen / Language.LSP.Types.TextDocumentDidClose messages to keep an in-memory filesystem of the current client workspace. The server can access and edit files in the client workspace by operating on the VFS in LspFuncs.

Documentation

data VFS

Constructors

VFS

Fields

vfsMap :: Map NormalizedUri VirtualFile

vfsTempDir :: FilePath

This is where all the temporary files will be written to

data VirtualFile

Constructors

VirtualFile

Fields

_lsp_version :: !Int

The LSP version of the document

_file_version :: !Int

This number is only incremented whilst the file remains in the map.

_text :: Rope

The full contents of the document

virtualFileText :: VirtualFile -> Text

virtualFileVersion :: VirtualFile -> Int

Managing the VFS

initVFS :: (VFS -> IO r) -> IO r

openVFS :: VFS -> Message 'TextDocumentDidOpen -> (VFS, [String])

Applies a DidChangeTextDocumentNotification to the VFS

Applies the changes from a DidOpenTextDocument to the VFS

changeFromClientVFS :: VFS -> Message 'TextDocumentDidChange -> (VFS, [String])

changeFromServerVFS :: VFS -> Message 'WorkspaceApplyEdit -> IO VFS

persistFileVFS :: VFS -> NormalizedUri -> Maybe (FilePath, IO ())

Write a virtual file to a temporary file if it exists in the VFS.

closeVFS :: VFS -> Message 'TextDocumentDidClose -> (VFS, [String])

updateVFS :: (VFSMap -> VFSMap) -> VFS -> VFS

manipulating the file contents

rangeLinesFromVfs :: VirtualFile -> Range -> Text

data PosPrefixInfo

Describes the line at the current cursor position

Constructors

PosPrefixInfo

Fields

fullLine :: Text

The full contents of the line the cursor is at

prefixModule :: Text

If any, the module name that was typed right before the cursor position. For example, if the user has typed "Data.Maybe.from", then this property will be Data.Maybe

prefixText :: Text

The word right before the cursor position, after removing the module part. For example if the user has typed "Data.Maybe.from", then this property will be "from"

cursorPos :: Position

The cursor position

getCompletionPrefix :: Monad m => Position -> VirtualFile -> m (Maybe PosPrefixInfo)

for tests

applyChanges :: Rope -> [TextDocumentContentChangeEvent] -> Rope

Apply the list of changes. Changes should be applied in the order that they are received from the client.

applyChange :: Rope -> TextDocumentContentChangeEvent -> Rope

changeChars :: Rope -> Int -> Int -> Text -> Rope